Fixed merge conflicts, still weird bug importing BASE in homescreen.js
