--- frontend/package.json
@@ -1,63 +1,0 @@
-{
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject",
-    "test": "jest --watchAll",
-    "lint": "eslint ./screens ./components ; exit 0",
-    "lint-fix": "eslint ./screens ./components --fix ; exit 0"
-  },
-  "eslintConfig": {
-    "extends": "universe/native"
-  },
-  "jest": {
-    "preset": "jest-expo"
-  },
-  "dependencies": {
-    "@babel/core": "^7.0.0",
-    "@expo/samples": "~3.0.3",
-    "@expo/vector-icons": "^10.0.3",
-    "@react-navigation/core": "^3.0.0-alpha.13",
-    "@react-navigation/web": "^1.0.0-alpha.9",
-    "Base64": "^1.1.0",
-    "airtable": "^0.7.1",
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "eslint-plugin-react-native": "^3.8.1",
-    "expo": "^35.0.0",
-    "expo-asset": "^7.0.0",
-    "expo-camera": "^7.0.0",
-    "expo-constants": "^7.0.0",
-    "expo-file-system": "^7.0.0",
-    "expo-font": "^7.0.0",
-    "expo-web-browser": "^7.0.0",
-    "react": "16.8.3",
-    "react-dom": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
-    "react-native-elements": "^1.2.6",
-    "react-native-gesture-handler": "~1.3.0",
-    "react-native-phone-input": "^0.2.4",
-    "react-native-screens": "^1.0.0-alpha.23",
-    "react-native-web": "^0.11.7",
-    "react-navigation": "^3.12.0",
-    "react-phone-number-input": "^2.3.24",
-    "styled-components": "^4.4.0",
-    "validate.js": "^0.13.1"
-  },
-  "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "babel-preset-expo": "^7.0.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.14.3",
-    "jest-expo": "^35.0.0",
-    "prettier": "^1.18.2"
-  },
-  "private": true
-}
