+++ lib/common.js
@@ -10,7 +10,6 @@ import getEnvVars from '../environment';
  * */
 
 const { AIRTABLE_API_KEY, AIRTABLE_BASE_ID, IMG_API_KEY } = getEnvVars();
-
 export default BASE = new Airtable({ apiKey: AIRTABLE_API_KEY }).base(
   AIRTABLE_BASE_ID
 );
